// src/types/IStatistical.ts
export interface IStatistical {
    id: number;
    sku?: string;
    productId: number;
    productName?: string;
    colorId?: number;
    colorCode?: string;
    colorName?: string;
    sizeId?: number;
    sizeName?: string;
    quantityInStock?: number;
    imageUrl?: string;
    totalQuantitySold?: number;
}